requirements.txt
alabaster==0.7.10
appdirs==1.4.3
Babel==2.4.0
chardet==2.3.0
dice==1.1.0
docopt==0.6.2
docutils==0.13.1
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
marshmallow==2.13.4
numpy==1.12.0
packaging==16.8
pdfminer.six==20160614
psutil==5.2.1
py==1.4.33
pydot==1.2.3
Pygments==2.2.0
pyparsing==2.2.0
PyQt5==5.8.2
pytest==3.0.7
pytest-mock==1.6.0
python-dateutil==2.6.0
python-graph-core==1.8.2
python-graph-dot==1.8.2
python-magic==0.4.13
pytz==2017.2
PyYAML==3.12
requests==2.13.0
sip==4.19.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.5
sphinx-paramlinks==0.3.2
